Add outputs for CrystalCast Spreadsheet.  Currently England, Scotland and Agegroup.  I will do English 7-regions next.
Agegroup is calculated as "R", but this comes from the rates that the group get infected, not the rates that they infect!

This R-output is using a 4-band smoothing filter to get today's R value.  Multiple other R-estimates still exist and are plotted.

Ask from JBC is to do an R value for each day. Will need to think how to do that!